Moved Mediafile class into `mediafile.py` to make it easier to disgusting what is exported in the __init__.py file.
